requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.4.2
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
toml==0.10.2
Werkzeug==2.0.1
mysqlclient==2.0.3
uwsgi